First pass at removing old message code
