--- frontstage/templates/secure-messages/message-form.html
@@ -1,45 +1,0 @@
-{% set message = {} if not message else message %}
-
-{% if not message %}<h1 class="saturn">Create message</h1>{% endif %}
-
-<div class="secure-message-form">
-    <form action="{{ url_for('secure_message_bp.create_message', ru_ref=ru_ref, survey=survey, case_id=case_id) }}" method="post">
-        <input type="hidden" name="csrf_token" value="{{ csrf_token() }}"/>
-        {{ form.thread_id }}
-        {{ form.thread_message_id }}
-
-        {% if message %}
-            {{ form.hidden_subject }}
-
-        {% else %}
-            <p>To: ONS Business Surveys team</p>
-            {% if "subject" in errors %}
-                <div class="panel panel--simple panel--error panel--spacious" id="subject-error">
-                    <p class="error-message">{{ errors['subject'][0] }}</p>
-            {% endif %}
-                <label class="venus" for="subject">Subject</label>
-                {{ form.subject(id='secure-message-subject', class_='secure-message-form__subject input input--text') }}
-            {% if "subject" in errors %}
-                </div>
-            {% endif %}
-            <br />
-        {% endif %}
-
-        {% if "body" in errors %}
-            <div class="panel panel--simple panel--error panel--spacious" id="body-error">
-                <p class="error-message">{{ errors['body'][0] }}</p>
-        {% endif %}
-        {% if message %}
-            <label class="venus" for="body">Reply</label>
-        {% else %}
-            <label class="venus" for="body">Message</label>
-        {% endif %}
-        {{ form.body(id='secure-message-body', class_='input input--textarea input--textarea-message', rows='10') }}
-        {% if "body" in errors %}
-            </div>
-        {% endif %}
-
-        <br />
-        {{ form.send(class_='btn' ) }}
-    </form>
-</div>
