+++ frontstage/templates/secure-messages/secure-messages-view.html
@@ -8,18 +8,6 @@
 
 <div class="secure-message-view-component">
 
-{% if label == 'DRAFT' and errors|length == 0%}
-    <div class="system-feedback"> 
-        <ul class="system-feedback__list"> 
-            <li class="system-feedback__list-item system-feedback__list-item--success"> 
-                <i class="fa fa-check-circle system-feedback__icon" aria-hidden="true" id="inbox-link"> </i>
-                Draft saved
-                <a href="{{ url_for('secure_message_bp.messages_get') }}">View your inbox</a> 
-            </li> 
-        </ul> 
-    </div>
-{% endif %}
-
     {% if message %}
         {% include 'secure-messages/thread-message.html' %}
     {% endif %}
@@ -28,12 +16,57 @@
         <div class="grid__col col-4@m"></div>
         <div class="grid__col col-7@m pull-1@m">
         {% if label != 'SENT' %}
-            {% include 'secure-messages/message-form.html' %}
+            {% set message = {} if not message else message %}
+
+            {% if not message %}<h1 class="saturn">Create message</h1>{% endif %}
+
+            <div class="secure-message-form">
+                <form action="{{ url_for('secure_message_bp.create_message', ru_ref=ru_ref, survey=survey, case_id=case_id) }}" method="post">
+                    <input type="hidden" name="csrf_token" value="{{ csrf_token() }}"/>
+                    {{ form.thread_id }}
+                    {{ form.thread_message_id }}
+
+                    {% if message %}
+                        {{ form.hidden_subject }}
+
+                    {% else %}
+                        <p>To: ONS Business Surveys team</p>
+                        {% if "subject" in errors %}
+                            <div class="panel panel--simple panel--error panel--spacious" id="subject-error">
+                                <p class="error-message">{{ errors['subject'][0] }}</p>
+                        {% endif %}
+                            <label class="venus" for="subject">Subject</label>
+                            {{ form.subject(id='secure-message-subject', class_='secure-message-form__subject input input--text') }}
+                        {% if "subject" in errors %}
+                            </div>
+                        {% endif %}
+                        <br />
+                    {% endif %}
+
+                    {% if "body" in errors %}
+                        <div class="panel panel--simple panel--error panel--spacious" id="body-error">
+                            <p class="error-message">{{ errors['body'][0] }}</p>
+                    {% endif %}
+                    {% if message %}
+                        <label class="venus" for="body">Reply</label>
+                    {% else %}
+                        <label class="venus" for="body">Message</label>
+                    {% endif %}
+                    {{ form.body(id='secure-message-body', class_='input input--textarea input--textarea-message', rows='10') }}
+                    {% if "body" in errors %}
+                        </div>
+                    {% endif %}
+
+                    <br />
+                    {{ form.send(class_='btn' ) }}
+                </form>
+            </div>
+
         {% endif %}
         </div>
     </div>
 
-    <p><a href="{{ url_for('secure_message_bp.messages_get') }}">Cancel</a></p>
+    <p><a href="{{ url_for('secure_message_bp.view_conversation_list') }}">Cancel</a></p>
 
 </div>
